backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
paho-mqtt==1.6.1
waitress==2.1.2
requests==2.31.0
scikit-learn>=1.0.0
numpy>=1.21.0
asyncua>=0.9.0
pymodbus>=3.0.0
asyncio